util: expose CheckBlockRWXVolumeForVM function
    - This function comes from checkOccupiedPVCs that we can reuse this
      logic to check volume on api level
